+++ crates/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # `moc-cli` Change Log
 
+## 0.5.3
+
+Realeased 2022-11-10
+
+* Add `TTYPE1=RANGE` keyword in FITS files (TTYPE is optional in the FITS standard but without
+  it astropy seems not to be able to read the file)
+* Add option `force_v1` to save a FITS file compatible with v1.0 of the MOC standard (S-MOC only).
+
 
 ## 0.5.2